requirements.txt
asgiref==3.8.1
attrs==24.2.0
bleach==6.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.16
django-debug-toolbar==4.4.2
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-recaptcha2==1.4.1
django-robots==6.1
django-simple-captcha==0.6.0
django-summernote==0.8.20.0
django-taggit==5.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
importlib_resources==6.4.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pillow==10.3.0
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.0
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
webencodings==0.5.1
